Изменения от заказчика

+++ Sketch1/ProgStruct.cpp
@@ -1,5 +1,5 @@
 struct PROG {   // Declare PERSON struct type  
-	int id;   // Declare member types  
-	int leng;
-	int amt;
+	long id;   // Declare member types  
+	long leng;
+	long amt;
 };
